tests/requirements.txt
beautifulsoup4==4.4.1
coverage==4.0.3
docutils==0.12
flake8==2.5.1
isort==4.2.2
natsort==4.0.4
pep8==1.7.0
py==1.4.31
pyflakes==1.0.0
Pygments==2.0.2
pytest-cov==2.2.1
pytest-django==2.9.1
pytest==2.8.5
sqlparse==0.1.18
